requirements.txt
crewai
langchain_openai
python-dotenv
crewai-tools
